requirements.txt
strands-agents==1.28.0
strands-agents-tools==0.2.21
defusedxml==0.7.1
youtube-transcript-api>=0.6.0
